// empala/aou-backend/products/core-be/src/enums/onboarding/unsupported-account-error-reason.ts
import { registerEnumType } from 'type-graphql';

export enum EUnsupportedAccountErrorReason {
  NON_CASH_ACCOUNT_TYPE = 'NON_CASH_ACCOUNT_TYPE',
  NON_DOMESTIC_ACCOUNT = 'NON_DOMESTIC_ACCOUNT',
  NON_US_CITIZEN = 'NON_US_CITIZEN',
  POLITICAL_OR_PUBLIC_PERSON = 'POLITICAL_OR_PUBLIC_PERSON',
  PUBLIC_COMPANY_10_PERCENT_SHARE_HOLDER = 'PUBLIC_COMPANY_10_PERCENT_SHARE_HOLDER',
  AFFILIATED_WITH_EXCHANGE_OR_FINRA = 'AFFILIATED_WITH_EXCHANGE_OR_FINRA',
}

registerEnumType(EUnsupportedAccountErrorReason, { name: 'EUnsupportedAccountErrorReason' });