// empala/aou-backend/products/core-be/src/enums/onboarding/account-validation-error-type.ts
import { registerEnumType } from 'type-graphql';

export enum EAccountValidationErrorType {
  INVALID_DATE_OF_BIRTH = 'INVALID_DATE_OF_BIRTH',
  TRUSTED_CONTACT_MISSING_FOR_55_OR_OLDER_USER = 'TRUSTED_CONTACT_MISSING_FOR_55_OR_OLDER_USER',
  MISSING_TRUSTED_CONTACT_FIELDS = 'MISSING_TRUSTED_CONTACT_FIELDS',
  MISSING_EMPLOYMENT_STATUS_FIELDS = 'MISSING_EMPLOYMENT_STATUS_FIELDS',
  MISSING_TRANSFER_ON_DEATH_FIELDS = 'MISSING_TRANSFER_ON_DEATH_FIELDS',
  MISSING_TRANSFER_ON_DEATH_AGREEMENT = 'MISSING_TRANSFER_ON_DEATH_AGREEMENT',
  MISSING_CUSTOMER_AGREEMENT = 'MISSING_CUSTOMER_AGREEMENT',
  MISSING_CUSTOMER_CUSTODIAN_AGREEMENT = 'MISSING_CUSTOMER_CUSTODIAN_AGREEMENT',
  MISSING_PRIMARY_APPLICANT_SIGNATURE = 'MISSING_PRIMARY_APPLICANT_SIGNATURE',
  MISSING_MAILING_ADDRESS_FIELDS = 'MISSING_MAILING_ADDRESS_FIELDS',
  MISSING_THIRD_PARTY_AUTHORIZED_TRADING_AGENT = 'MISSING_THIRD_PARTY_AUTHORIZED_TRADING_AGENT',
}

registerEnumType(EAccountValidationErrorType, { name: 'EAccountValidationErrorType' });